Extension interceptor
Here, we rely on the driver wrapping a call to CreateCookie and utilising
the WrapBeforeCall, WrapAfterCall and WrapClose methods to add behaviour
to the ContainerTask.

(The api/agent/drivers/stats package needs to be extracted since this
introduces a reference to *models.Call in `drivers`.)
